Refactor pathioc
